+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - change settings amount popup
+- spends storage
 
 ## [3.1.0] - 2021-12-24
 ### Added